requirements.txt
flask
Flask-SQLAlchemy
Flask-Mail
flask-marshmallow
bcrypt
jwt
marshmallow-sqlalchemy
marshmallow-sqlalchemy
flask_cors
psycopg2
flask-mail
parse
